Separate DDP and DMP pgs to improve T16 training QPS (#2133)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2133

In t16 hosts, we don't have backend network, the frontend network has longer latency which is not easy to optimize.
To train with multiple t16 nodes, we need to paralize comm as much as possible, to achieve the best performance.
This change allows sparse and dense comm using different PG with NCCL, so we can issue network request concurrently without pending on each other.
There is no dependency between sparse and dense result, except the overarch computation, so there is no risk to model training result.

Reviewed By: goldcoderZ

Differential Revision: D58691294