Update criteo.py
The original script simply added 3 to the target value before taking the log. This led to the issue that in data preprocessing, if there was a value of -3, it would result in a value of -inf. This problem was mentioned in the issue https://github.com/facebookresearch/dlrm/issues/363#issue-1963837683. I changed the preprocessing operation to dense_np -= (dense_np.min() - 2) in the tsv_to_npys function, and correctly handled the Criteo Kaggle dataset.