Call finish_checkpoint_resume to flush optimizer step (#2244)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2244

X-link: https://github.com/pytorch/FBGEMM/pull/2884

`iter` number is a critical counter in computing how much decay is needed together with `prev_iter`. However, `iter` is not checkpointed since it is not in `split_optimizer_state`. Instead, it relies on external calls of `set_optimizer_step` to flush the counter, which relies on explict calls of `finish_checkpoint_resume` to trigger `flush_state` in `FullSyncOptimizer`. This chain of actions is not properly set up in existing code, which causes the issue that after a preemption, `iter` gets reset to zero.

Reviewed By: yuxihu

Differential Revision: D60155781

fbshipit-source-id: 51b85fe10fb196df52bfdfc22424e6ab6f8c5bfb