fix torch rec test failure
Summary:
Fixes T192448049. The module call form an unusal call stack for the nodes: https://www.internalfb.com/phabricator/paste/view/P1507230978. This is currently not supported by unflattener and need some extra design to make it work. We'll comment it out for now.

A TODO is also added to the code-base of unflattener in D60528900

Differential Revision: D60682384