Support global norm gradient clipping if DTensor is used (#2271)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2271

If DTensor is used in the parameters passed to gradient clipping, we compute the global norm across all ranks (instead per-rank local norm) to be clipped. This is needed according to a study in https://fb.workplace.com/chat/t/26249100244704391#:~:text=was%20really%20important%3A-,D59763695,-today%20in%20APS

The existing test for gradient clipping is modified to test this new capability.

Note: the global norm gradient clipping was original implemented by Andrew in D60625965. I combine it with the unit test here as single diff.

Differential Revision: D60704597