Fix device issue with regroup module (#2253)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2253

When using the new RegroupAsDict module with cached tensors, ex. self._permutes, they aren't properly moved to the right devices during the model processing pipeline. This is due to the fact that the module is not a leaf module and does not have a `set_device` method.

This diff creates a submodule `PermutePooledEmbeddings` that just handles the fbgemm logic and holds the cached tensors (for moving to the right device), along with adding this module to the leaf modules for AIMP.

Reviewed By: TroyGarden

Differential Revision: D60062406