remove _embedding_collection alias in mc_embedding that causes extra ckpt storage (#2290)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2290

the alias will make checkpoint create extra files to store the values (although having the same value). checked the code and dstaay-fb it is not necessary. The _embedding_collection style is mainly used when parent is sparse_arch, which does not seem to be our case either. so removing it.

Reviewed By: dstaay-fb, bixue2010, sarckk

Differential Revision: D61149486

fbshipit-source-id: 7aae56ee00889ef05fa0cefd93734d303990185c