Add a DpProposer to enhance collective_plan execution speed (#2286)
Summary:
The collective_plan will encounter an NCCL timeout if the number of EmbeddingBagConfig exceeds 500, as noted in https://github.com/pytorch/torchrec/issues/2282 , To address this issue, we propose using a Proposer with dynamic programming to improve the execution speed of the collective plan.

In the case of `test_multiple_ebc_plan.py` mentioned in https://github.com/pytorch/torchrec/issues/2282, the collective plan with the DpProposer took only 11.86 seconds, whereas the original version took 904.90 seconds.

Pull Request resolved: https://github.com/pytorch/torchrec/pull/2286

Reviewed By: PaulZhang12

Differential Revision: D61250676

Pulled By: sarckk

fbshipit-source-id: 3a158f6e19a677b8558faed57d57e738f81c1fde