Consisent indices handling (#2330)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2330

So:

(1) we did not have consistent handling between quant and sharded quant dtypes.
(2) users correctly identified on cpu path we want to make sure indices and offsets are same dtype (to save templates).  But talking to sryap we want to just make them consistent, and leave in int32 if possible.
(3) cuda is always int32. this is likely to be issue with ZCH models pushing beyond 2B row shards for inference, for now just need make shards that size.

Reviewed By: sryap

Differential Revision: D61446242

fbshipit-source-id: 779d22822d4ac7656675d9ac3887b5ac3ecbfa39