Do not ignore `None` arg when pipelining (#2352)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2352

if `arg` to embedding module is `None`, we would ignore it. However, now we also use `_get_node_args_helper` to generate arg list info for preproc modules, and sometimes `None` is passed in as arg/kwarg.

With changes in https://github.com/pytorch/torchrec/pull/2342, we can now handle constants. For backward compatibility, adding an optional flag to indicate to `_get_node_args_helper` that we are handling preproc modules.

Reviewed By: xing-liu

Differential Revision: D61938346

fbshipit-source-id: 1514fb0432be10b7fa46f90050ecbe370cf07a4c