Remove Option for ProcessGroup and Expose backend Options to reflect the correct code structure (#132931) (#2384)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2384

X-link: https://github.com/pytorch/pytorch/pull/135653

We introduced the dispatchable backend for a ProcessGroup and collective in https://github.com/pytorch/pytorch/issues/86225. This PR is a follow-up cleanup to clean up the option of a ProcessGroup and ask users to either set timeout or backend later on or directly create backend after creating a PG.

Also PGNCCL is using option class from ProcessGroup but we actually should use Option from backend class. So this PR is to make the type or name to be aligned with what we are doing in cpp side. I don't change the signature for the public API, so they still use args named "pg_options"

We need to make changes to the test to make it aligned with the change.

This is try to reland D62008954 by fixing internal errors.
ghstack-source-id: 242088446

Reviewed By: wz337, H-Huang

Differential Revision: D62483294

fbshipit-source-id: c31184311a0f72044f506e75a78d5dac0471bc41