add GPU sync and numerical value tests (#2194)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2194

Added GPU sync tests to simulate gathering metric states on to rank 0 and computing. Tests don't cover this case before, which has resulted in SEVs in the past as users aren't aware of how RecMetrics collects and computes metrics.

Added numerical value tests, most metrics are do not have this which can result in issues down the line if metrics need to be changed/accomodate future changes. Also we've found inconsistencies sometimes from other methods, so always good to check here. We compare each metric to a reference implementation from literature to ensure the values are as expected.

Reviewed By: henrylhtsang

Differential Revision: D59173140