Update default value for TorchCompileConfig to be None (#2433)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2433

If we set dynamics to False, then automatic_dynamic_shapes will not take effect. This is not ideal because for sparse arch compilation, we expect that there are a number of dynamic shapes, if we do not apply automatic_dynamic_shapes, compiler will quit compilation after 8 trials.

Thus, we change it to None in this diff which is also the default setting for `torch.compile`:

https://www.internalfb.com/code/fbsource/[08c8192803b22f831329c8f736a3f7c6093ea4a8]/fbcode/caffe2/torch/__init__.py?lines=2321

Reviewed By: Yuzhen11

Differential Revision: D63476875