grid based sharding for EBC (#2445)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2445

## Introducing Grid based sharding for EBC
This is a form of CW sharding and then TWRW sharding the respective CW shards.

One of the key changes is how the metadata from sharding placements is constructed in grid sharding. We leverage the concept of `per_node` from TWRW and combine it with the permutations and concatenation required in CW.

#### Comms Design
Combining optimizations from both CW and TWRW to have highly performant code. Sparse feature distribution is done akin to the TWRW process where the input KJT is bucketized and permuted (according to the stagger indices) and AlltoAll to the correponding ranks that require their part of the input.

Similarly for the embedding all to all, we first all reduce within the node to get the embedding lookup for the CW shard on that node. We're able to leverage the intra node comms for this stage.  Then the reduce scatter, just like TWRW, where the embeddings are split to each rank to hold it's and its corresponding cross group ranks embeddings which are then shared in a AlltoAll call. Lastly, the `PermutePooledEmbeddingsSplit` callback is called to rearrange the embedding lookup appropriately (cat the CW lookups in the right order).

#### Optimizer Sharding
Fused optimizer sharding is also updated, we needed to fix how row wise optimizer states are constructed since optimizers for CW shards are row wise sharded. For grid sharding, this doesn't work since the row wise shards are repeated for each CW shard as well as we can encounter the uneven row wise case which is not possible in CW sharding. For grid shards the approach is to use a rolling offset from the previous shard which solves for both uneven row wise shards and the repeated CW shards.

NOTE: Bypasses added in planner to pass CI, which are to be removed in forthcoming diff

Reviewed By: dstaay-fb

Differential Revision: D62594442