Enable default TW sharding for modules in AIMP
Summary:
D62780756 exposed a bug where the `shard_quant_model` API only defaults to TW sharding for EBC and no other module types - this was hard coded.

This diff will enable TW sharding as default sharding for all modules, unblocking RecGPT case where we will need VLE module to default to TW sharding.

We decided to hardcode on AIMP side to specify constraints for default TW sharding - because VLE is internal only so should not be exposed in OS code.

Reviewed By: PaulZhang12

Differential Revision: D63552253