Handle train/eval mode in base class (#2464)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2464

When running for evaluation, the model needs to be set to eval mode by setting self.training=False. This flag should be propagated to all sub-modules. However, for `ShardedEmbeddingModule` only the weights of TBE are registered (https://fburl.com/code/9g0uykxu) but the TBE module is not taken care of. This causes a problem when running eval: the TBE is not put into eval mode, especially when GWD is enabled which requires different behavior for training vs eval (eval shouldn't be doing weight decay).

D59817443 supported ShardedPooledEmbeddingArch. Here we upstream the support to the base class `ShardedEmbeddingModule` so that all other impl can directly inherit.

Reviewed By: dstaay-fb

Differential Revision: D63758562