Integrate PT2 pipeline with FullModelCompilation Config (#2475)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2475

In current FMC setting, we start compilation from the first iteration. However, when we do sparse arch compilation, torchrec needs a few iterations of dry runs before compilation. So we need to postpone the compilation to happen in the torchrec pipeline.

Despite the difference, we aim to provide an uniform interface for users. Users can set `FullModelCompilation Config` as usual and if users also use `training.pipeline_type=pt2`, this diff can handle the postpone mechanism automatically.

Reviewed By: Yuzhen11

Differential Revision: D63500331

fbshipit-source-id: 0291c43fed6fb7fc555bc3b4524f9a9f79a13c45