Create an interface to enable eviction policy (#2479)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2479

To support various types of eviction policy, the `HashZchManagedCollisionModule` needs to be able to calculate a score (e.g., TTL) for each incoming ID and pass it to the kernel. The latter will make an informed eviction decision based on the existing score associated with the identity (stored in metadata), and a reference value (e.g., current timestamp). Refer to [Doc](https://docs.google.com/document/d/19s6N4-CWhveqhV1ePwrHAF8SIlzePsoxKElPRnnlo8c/edit?usp=sharing) for details.

This Diff uses per-feature TTL as an example, and creates a scorer to assign a TTL for each ID with regard to the expiration hr of the feature. The score will be passed to the kernel (in a separate Diff) to perform eviction.

Differential Revision: D63879449