Create an interface to enable eviction policy (TREC part) (#2481)
Summary:

To support various types of eviction policy, the HashZchManagedCollisionModule needs to be able to calculate a score (e.g., TTL) for each incoming ID and pass it to the kernel. The latter will make an informed eviction decision based on the existing score associated with the identity (stored in metadata), and a reference value (e.g., current timestamp).

Differential Revision: D64163927