- For planner to use cpu for search random generator (#2484)
Summary:

torch.rand() defaults to using the default device. If torch.device has
been globally set to 'meta', then this breaks the planner code. Force
the device to cpu instead. This ensures we're using the same seeded
random number alg as before (so no changes to existing plans). We only
generate a handful of random numbers so performance is not a concern.

Example breakage:
https://fb.workplace.com/groups/260102303573409/posts/507200532196917/

Differential Revision: D64260019