Create an interface to enable eviction policy (TREC part) (#2481)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2481

To support various types of eviction policy, the HashZchManagedCollisionModule needs to be able to calculate a score (e.g., TTL) for each incoming ID and pass it to the kernel. The latter will make an informed eviction decision based on the existing score associated with the identity (stored in metadata), and a reference value (e.g., current timestamp).

Reviewed By: dracifer, shruthign

Differential Revision: D64163927

fbshipit-source-id: b5e65fbf0568e7a4950988128fd19c1ffa4426ca