make it configurable to use flattened/unflattened length for rebatching (#2536)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2536

Previously we use _get_batching_hinted_output to infer batch infos for stbe output and length, and use flattened length for rebatching. This won't work for IG stories model as the model has stbe input with pooling factor > 1, so we have to use unflattened length for rebatching. To avoid affecting previous models that's already used flattened length, we add an option to choose which one to use.

Reviewed By: PaulZhang12

Differential Revision: D65442322

fbshipit-source-id: 4ec0d94821ebe6059a6a7d9d40c3e6ac9a053069