Favor sharding options with fused_uvm_caching kernel in (#2551)
Summary:

TorchRec auto-planner doesn't work well with embedding offloading enabled. With auto-planner enabled the number of sharding options in the search space is too large to explore and `EmbeddingoffloadScaleupProposer` rarely selects a sharding option with `fused_uvm_caching`.

An alternative approach to D65162749 - instead of adding constraints in the trainer code, this diff filters out and selects sharding options in EmbeddingOffloadScaleupProposer that have the `fused_uvm_caching` compute_kernel if possible. If no sharding options fufill this condition, we choose the best scoring sharding option (as before).

We still only look at one sharding option so there is a follow up to do more investigation on improving the proposer: T206831945.

Reviewed By: damianr99

Differential Revision: D65774286