Remove QuantizeEBC grouping by data type
Summary:
FBGEMM supports TBEs being initialized with a list of data_types, so we no longer have to split tables with different quantization data types into separate tables.

We've already implemented this for the ShardedQuantizedEBC - see base diff - which helped to optimize TorchRec eager mode Inference to have on-par QPS with non-eager mode inference.

This diff introduces the same optimization to QuantizedEBC - even though it's not needed, for consistency.

Differential Revision: D63861064