Remove QuantizeEBC grouping by data type (#2571)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2571

FBGEMM supports TBEs being initialized with a list of data_types, so we no longer have to split tables with different quantization data types into separate tables.

We've already implemented this for the ShardedQuantizedEBC - see base diff - which helped to optimize TorchRec eager mode Inference to have on-par QPS with non-eager mode inference.

This diff introduces the same optimization to QuantizedEBC - even though it's not needed, for consistency.

Reviewed By: PaulZhang12

Differential Revision: D63861064

fbshipit-source-id: 376adbce6cde5d9c45def0836c78c1455993d419