Remove unused-variable in tiles/test/TilesTestBase.cpp +3 (#2572)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2572

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: wuyuoss

Differential Revision: D66143538

fbshipit-source-id: 18bf044ca416517b78669f86cc43f9ba0a6e3964