Narrow scaleup probes to max cache sharding span (#2588)
Summary:

When a lot of cache scaleup budget is available, significantly larger
than the total amount of memory needed to promote every table to HBM,
it's likely that many of the budget probes will attempt to cost a plan
using more budget than the proposal can utilize. In these scenarios,
we tend to see only two distinct plan costs, 1) the min-working-set
plan which is costed first, 2) every other proposal "clips" at the max
scaleup limit (i.e. everything promoted to HBM).

It's also plausible that the fully-promoted plan is more expensive
than the min-working-set plan, due to the increased bin-packing
difficulty of fitting the larger shards. In these cases, the job only
runs on the min-working-set proposal, even though (lots of) additional
memory is available for larger caches (up to the point of diminishing
returns due to bin-packing overhead).

This diff narrows the search region, when more memory is available
than we can use, to focus our search effort on productive portions of
the search space. This increases the likelihood we discover a plan
that is both cheaper than min-working-set or fully-promoted.

Reviewed By: keyan

Differential Revision: D66419942