Unify InferRwSequenceEmbedding Modules for GPU / CPU (#2559)
Summary:

Unify InferRwSequenceEmbedding Modules for GPU / CPU.

There does not seem to be much difference in the implementation for InferRwSequenceEmbedding and InferCPURwSequenceEmbedding.

For heterogeneous sharding, we need to merge them together into one module. 

Also introduced the concept of device_type_from_sharding_info to propagate the correct device for output dist.

Reviewed By: jiayisuse

Differential Revision: D65859663