add buck target for torchrec/inference/inference_legacy and suppress Pyre error
Summary: Internal and external Pyre is out of sync. One of the reasons is because one of the files is not currently being typechecked by Pyre due to not being included in a buck target. This diff adds a target to include this file so that Pyre will pick it up when typechecking, along with suppressing the actual error with a pyre-fixme

Reviewed By: PaulZhang12

Differential Revision: D67035607