explicitly call update_hyper_parameters (#2663)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2663

Due to various wrappers classes, the hyperparameters are not properly updated within the underlying optimizer, especially the ones that are not saved in `param_groups`. Therefore, we'd need to explicitly call the `update_hyper_parameters` method in order to channel the schedule and change the actual values used within the optimizer.

Reviewed By: xinzhang-nac

Differential Revision: D67804589

fbshipit-source-id: 4b4023b5187dd4783012747059a50fa70b0bda7a