fix ZCH inference input dist error (#2682)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2682

Serving test failed when ZCH sharding is enabled because embedding lookup request Out of Bound.

After compared with training sharding input_dist, found that the keep_orig_idx flag is True when calling training block bucketize kernel, but not set in inference. So this diff is to make that consistent between training and inference, so the ID distribution works the same between training and inference.

Reviewed By: kausv

Differential Revision: D68123290

fbshipit-source-id: 57813e55fc946dec79a522cfb341974a2ed7669d