Remove unused memory checks to speed up `compute` (#2719)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2719

The memory checks here have non-significant overhead in every compute step as there are a lot of tensor size calls involved here.
In our runs, this accounted for around 20% time spent in the rec metric compute step.

Given that this is not being used anymore, let's remove this call.

This diff removes the call from the metric_module. In the next set of diffs, I'll remove the argument from the callsites.

Reviewed By: fegin

Differential Revision: D68995122

fbshipit-source-id: 34e9d2399c3c647c41585a3f0efba2ad404cc8f4