Make dict update fx compatible
Summary:
Dict's `update` method cannot be supported by *fx*:
```
NotImplementedError: 'immutable_dict' object does not support mutation. If you are attempting to modify the kwargs or args of a torch.fx.Node object,
```

It needs to be handled for every key-value pair.

Not 100% sure why this issue was not surfaced before. But likely it is because we have never published a model with more than one MPZCH-powered embedding table, so the code path was not hit before.

Differential Revision: D70227971