Fix Linux wheels builds for Python 3.9 + cpu (#2806)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2806

This has been failing for a while - for some reason `importlib-metadata` is not installed for the linux cpu builds. While I investigate specifically why this is an easy way to unblock and get those jobs working again.

Also renamed the file since we are not just installing fbgemm anymore.

Reviewed By: iamzainhuda

Differential Revision: D71006875

fbshipit-source-id: 2e0eb7c94434e7aa5c970ff0061ff733ee686ac3