Add option to skip certain unit tests (#2878)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2878

Add option to skip certain tests - useful when a few tests are broken and we need to skip them to gain visibility in other unit tests. This usually happens on our CPU unit tests, so only modifying this script.

To use, simply add the name of the tests to skip in this txt file. such as:

```
test_sharding_fused_ebc_as_top_level
```

You can also use the class name for the test: e.g.
```
ModelParallelSparseOnlyTestGloo
```

Differential Revision: D72815908