Perf Model Modeling Critical Path of Sparse Arch (#2896)
Summary:

Update the critical path definition in the planner logs to match what we think is the most realistic option. See the docstring for the detailed modeling.

The goal is to later feed this into the proposer and partitioner to test if we can improve upon the current greedy algorithm used with the new perf model. 

We will also replace a [similar version](https://www.internalfb.com/code/fbsource/[50e47c413eb3d4e00facb9df71592f9aa81d8aee]/fbcode/torchrec/distributed/planner/stats.py?lines=1071) of this function in the stats.py to avoid duplicating the logic.

Differential Revision: D73207877