Add unsharded module property to sharded modules (#2901)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2901

Adding a simple unsharded module reference to sharded modules. This will be used in Dynamic Sharding by `DistributedModelParallel` to reshard an already-sharded_module.

As DMP is created with only one-way relationship in mind, accessing the unsharded module type will help determine which sharder to use in 'resharding'.

Most of the changes here are simply to add in the property in where ShardedModule or it's wrapper ShardedEmbeddingModule is used.

Differential Revision: D73407830