Enable More General Comms Bandwidth Estimates (#2905)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2905

Optimizing sharding plans relies on good perf estimates which in turn relies on good bandwidth estimates.

Bandwidth of a collective communication depends on many factors including the world size, type of collective (all to all, reduce scatter, ...), and hardware.

In this change, we introduce an abstract class GeneralizedCommsBandwidths that implements a get_bw function. Users can subclass this and define the get_bw appropriately for the hardware or network setup being used. If passed to the Topology, it will be called from perf estimators and will enable setting correct configurations of bandwidth depending on hardware, collective, world_size. We prefer to define this as a class and function as opposed to a dictionary or other similar data structure because we want to enable users to specify bandwidth as precisely as is useful to them. For example, specifying a different bandwidth for every different worldsize rather than only <local_world_size or >=local_world_size.

For Topology objects, the comms_bandwidths argument is optional so we do not change existing behavior and perf estimates do not change. If not passed, we create a GeneralizedCommsBandwidth object with the passed values for inter and intra node comms bw and that replicates the existing behavior of perf estimators.  We re-define the signature of perf estimators downstream to ensure consistency and clarity for users so that there is a consistent comms bandwidth.

Reviewed By: iamzainhuda

Differential Revision: D73229318

fbshipit-source-id: 5dd1e6066ea3686a944be4da4b6ac6a97597cdfc