add activation on last logic
Summary:
# context

Adding parity for activation on last flag to make module more configurable.

activation_on_last is a flag that toggles the given (or default) activation function to the last layer of the MLP. Typically ALL layers of the MLP will have the activation function. If it is set to false, then the last layer will not have the activation function applied. This is so users can optionally use the raw MLP output for their own customized needs.

Reviewed By: TroyGarden

Differential Revision: D73691616